requirements.txt
flask==3.0.0
python-dotenv==1.0.0
openai==1.12.0
Pillow==10.2.0 